requirements.txt
flask
openai
sentence-transformers
faiss-cpu
